refactor(muya): move block-transform logic out of UI plugins into the engine
The core engine (`muya.ts`) imported block-transformation and turn-into
logic from inside two UI-plugin folders — a reversed dependency (engine →
src/ui). That logic is engine behavior, not UI, so it now lives in a new
engine module and both the engine and the UI plugins depend on it.

- Add `src/block/blockTransforms.ts` holding the document-editing helpers
  moved verbatim out of `ui/paragraphQuickInsertMenu/config.ts`:
  `frontmatterMeta`, `insertFrontMatterAtStart`, `buildReplacementBlock`
  (+ block builders), `replaceBlockByLabel`, `insertBlockBelowByLabel`,
  `showTablePicker`, and the caret helpers.
- Extract a pure `canTurnInto(block, label)` predicate (label regexes ported
  verbatim) into the new module; `canTurnIntoMenu` in
  `ui/paragraphFrontMenu/config.ts` is now a thin UI wrapper that filters the
  menu items through it, so the front menu keeps its icon-rich output while
  the engine consumes only the predicate.
- Repoint `muya.ts` and both paragraph plugins at the new module; prune the
  now-unused imports left behind in `paragraphQuickInsertMenu/config.ts`
  (which keeps only `MENU_CONFIG`, `IQuickInsertMenuItem`, `getLabelFromEvent`).
- Update the two specs that imported `replaceBlockByLabel` directly.

No behavior change: the moved functions are unchanged and `canTurnInto`
reproduces the previous filtering exactly.
