+++ packages/muya/src/ui/paragraphFrontMenu/config.ts
@@ -5,6 +5,7 @@ import type {
 import copyIcon from '../../assets/icons/copy/2.png';
 import deleteIcon from '../../assets/icons/delete/2.png';
 import newIcon from '../../assets/icons/paragraph/2.png';
+import { canTurnInto } from '../../block/blockTransforms';
 import { isOsx } from '../../config';
 import {
     MENU_CONFIG,
@@ -41,39 +42,5 @@ export const FRONT_MENU = [
 export type FrontMenuIcon = (typeof FRONT_MENU)[number];
 
 export function canTurnIntoMenu(block: Parent) {
-    const { blockName } = block;
-
-    switch (blockName) {
-        case 'paragraph': {
-            const paragraphIsEmpty = /^\s*$/.test(block.firstContentInDescendant()!.text);
-            if (paragraphIsEmpty)
-                return ALL_MENU_CONFIG.filter(item => item.label !== 'frontmatter');
-
-            const PARAGRAPH_TURN_INTO_REG
-                = /paragraph|atx-heading|block-quote|order-list|bullet-list|task-list/;
-
-            return ALL_MENU_CONFIG.filter(item =>
-                PARAGRAPH_TURN_INTO_REG.test(item.label),
-            );
-        }
-
-        case 'atx-heading': {
-            return ALL_MENU_CONFIG.filter(item =>
-                /atx-heading|paragraph/.test(item.label),
-            );
-        }
-
-        case 'order-list':
-            // fall through
-        case 'bullet-list':
-            // fall through
-        case 'task-list': {
-            return ALL_MENU_CONFIG.filter(item =>
-                /order-list|bullet-list|task-list/.test(item.label),
-            );
-        }
-
-        default:
-            return [];
-    }
+    return ALL_MENU_CONFIG.filter(item => canTurnInto(block, item.label));
 }
